style: 拆分原sstable.rs和version.rs组件
