+++ src/kernel/lsm/log.rs
@@ -9,7 +9,7 @@ use std::sync::Arc;
 use integer_encoding::FixedInt;
 use crate::kernel::{Result, sorted_gen_list};
 use crate::kernel::io::{FileExtension, IoFactory, IoType, IoWriter};
-use crate::kernel::lsm::lsm_kv::Gen;
+use crate::kernel::lsm::storage::Gen;
 use crate::KernelError;
 
 const BLOCK_SIZE: usize = 32 * 1024;
@@ -278,8 +278,8 @@ mod tests {
     use crate::kernel::io::IoType;
     use crate::kernel::lsm::log::{HEADER_SIZE, LogLoader, LogReader, LogWriter};
     use crate::kernel::Result;
-    use crate::kernel::lsm::lsm_kv::Config;
     use crate::kernel::lsm::mem_table::DEFAULT_WAL_PATH;
+    use crate::kernel::lsm::storage::Config;
 
     #[test]
     fn test_writer() {
